pyproject.toml
[project]
name = "playwright-changesets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "playwright==1.51.0",
    "polars>=1.29.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pip>=25.1",
]
